// src/types.ts
export enum ThemeParam {
  breakpoints = 'breakpoints',
  fontSizes = 'fontSizes',
  fontWeights = 'fontWeights',
  letterSpacings = 'letterSpacings',
  lineHeights = 'lineHeights',
  radii = 'radii',
  sizes = 'sizes',
  space = 'space',
  zIndices = 'zIndices',
}

export enum ValueUnits {
  px = 'px',
  em = 'em',
  rem = 'rem',
  percent = '%',
  none = '',
}